Moved to middleware, still goofy but bit better
formatting, linting
